requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
stripe
pydantic
pydantic[email]
python-multipart
cryptography
